cont. surah list screen
